Specify min/max value as string in error message
The fractional part is not visible in the error message of the following
methods:

>>> Assert\Assertion::min(0.1, 0.2);
Assert\InvalidArgumentException with message 'Number "0.1" was expected
to be at least "0".'

Should be: Number "0.1" was expected to be at least "0.2".

>>> Assert\Assertion::max(0.2, 0.1);
Assert\InvalidArgumentException with message 'Number "0.2" was expected
to be at most "0".'

Should be: Number "0.2" was expected to be at most "0.1".