LazyAssertion magic calls return $this (#278)
A LazyAssertion magic call needs to return $this.
If you extend the Assertion class with your own assertion methods,
It is great if that methods are also hinted in your IDE.

To do this, you can override the LazyAssertion class and add your own
phpDoc at class level. But if you chain the methods with existing
assertions, it will not work anymore because they return the
default LazyAsssertion object and not your own LazyAssertion object.