Remove a hard dependency on the intl extension
Fixes #288.

While #268 and #269 updated count() and isCountable() to now work with intl resources, #286 made that the whole beberlei/assert package now requires the intl extension to be installed.

However, the intl extension dependency is a soft dependency.
The package does not need intl for most of its features and can safely work without it.